move SafeCoreCrypto to new file
