+++ wire-ios-data-model/Source/Core Crypto/SafeCoreCrypto.swift
@@ -1,0 +1,92 @@
+//
+// Wire
+// Copyright (C) 2026 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import WireCoreCrypto
+import WireLogging
+import WireUtilitiesPackage
+
+/// A wrapper object for CoreCrypto that ensures transactions are executed
+/// within a background task (if a task manager is provided).
+public final class SafeCoreCrypto {
+
+    private let backgroundTaskManager: (any BackgroundTaskManager)?
+    let coreCrypto: any CoreCryptoProtocol
+
+    public init(
+        backgroundTaskManager: (any BackgroundTaskManager)?,
+        coreCrypto: any CoreCryptoProtocol
+    ) {
+        self.backgroundTaskManager = backgroundTaskManager
+        self.coreCrypto = coreCrypto
+    }
+
+    public func registerEpochObserver(_ epochObserver: any EpochObserver) async throws {
+        try await coreCrypto.registerEpochObserver(epochObserver)
+    }
+
+    public func transaction<Result>(
+        block: @escaping (any CoreCryptoContextProtocol) async throws -> Result
+    ) async throws -> Result {
+        if let backgroundTaskManager {
+            try await transaction(
+                backgroundTaskManager: backgroundTaskManager,
+                block: block
+            )
+        } else {
+            try await coreCrypto.transaction(block)
+        }
+    }
+
+    /// Perform a transaction within the context of a background task to
+    /// ensure the transaction has additional time to complete while the
+    /// app transitions to a suspended state.
+    ///
+    /// This is particularly important when running in the main app because
+    /// if the app is suspended during a transaction, then core crypto will
+    /// hold on to a file lock used to coordinate cross-process concurrency
+    /// and effectively block app extensions from being able to start their
+    /// own transactions. This can lead to the Notification Service Extension
+    /// being blocked and not process any notifications until the main app
+    /// is resumed and the transaction is completed and the file lock released.
+    private func transaction<Result>(
+        backgroundTaskManager: any BackgroundTaskManager,
+        block: @escaping (any CoreCryptoContextProtocol) async throws -> Result
+    ) async throws -> Result {
+        let transactionTask = Task {
+            try await coreCrypto.transaction(block)
+        }
+
+        let taskID = backgroundTaskManager.beginBackgroundTask(
+            withName: "core crypto transaction",
+            expirationHandler: {
+                WireLogger.coreCrypto.warn(
+                    "Background task expiring, cancelling core crypto transaction",
+                    attributes: .safePublic
+                )
+                transactionTask.cancel()
+            }
+        )
+
+        defer {
+            backgroundTaskManager.endBackgroundTask(taskID)
+        }
+
+        return try await transactionTask.value
+    }
+
+}
